Fix Test_1_1

+++ src/main/java/jmorfsdk/form/Form.java
@@ -78,7 +78,7 @@ public abstract class Form {
         return getMyControlValue() == getControlHashCode(string);
     }
 
-    private byte getMyControlValue() {
+    private int getMyControlValue() {
         return getControlValue(getMyFormKey());
     }
